+++ src/Redis/Client/CredisClient.php
@@ -51,13 +51,6 @@ class CredisClient extends AbstractClient
         return new self($redis);
     }
 
-    public function execute(Command $command)
-    {
-        $result = $this->redis->__call($command->getId(), $command->getArguments());
-
-        return $command->parseResponse($this->fixFalseToNull($result));
-    }
-
     public function executeRaw(...$args)
     {
         $command = array_shift($args);