+++ src/Redis/Client/PhpiredisClient.php
@@ -51,13 +51,6 @@ class PhpiredisClient extends AbstractClient
         $this->redis = $redis;
     }
 
-    public function execute(Command $command)
-    {
-        $rawResponse = phpiredis_command_bs($this->redis, array_merge([$command->getId()], $command->getArguments()));
-
-        return $command->parseResponse($rawResponse);
-    }
-
     public static function supports($redis): bool
     {
         if (!is_resource($redis)