+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add client for ReactPHP Redis
 - Path validation for JSON fields
 - Helper function get current RedisJSON version + (abstraction)
 - (dev) More unit tests